+++ packages/vue/src/components/index.ts
@@ -42,6 +42,16 @@ export type {
 export { ProgressBar, progressBarTheme } from './ProgressBar'
 export type { ProgressBarProps, ProgressBarTheme } from './ProgressBar'
 
+export { BigText, bigTextTheme } from './BigText'
+export type {
+	BigTextProps,
+	BigTextTheme,
+	BigTextFont,
+	BigTextAlign,
+	BigTextBackgroundColor,
+	BigTextGradient,
+	BigTextEnv,
+} from './BigText'
 export { ScrollView } from './ScrollView'
 export type { IScrollViewProps, IScrollViewHandle } from './ScrollView'
 //#endregion Display Components